lots of work on refactoring for history
