+++ examples/links/index.js
@@ -60,7 +60,6 @@ class Links extends React.Component {
 
   onChange = (state) => {
     this.setState({ state })
-    console.log(state.document.toJS())
   }
 
   /**
