+++ lib/models/transform.js
@@ -1,7 +1,5 @@
 
-import Operations from '../transforms/operations'
 import Transforms from '../transforms'
-import { List, Record } from 'immutable'
 
 /**
  * Transform.
@@ -18,10 +16,13 @@ class Transform {
    */
 
   constructor(properties) {
-    const { state } = properties
-    this.initialState = state
+    const { state, operations = [] } = properties
     this.state = state
     this.operations = []
+
+    operations.forEach(op => {
+      this.applyOperation(op)
+    })
   }
 
   /**
@@ -35,26 +36,6 @@ class Transform {
   }
 
   /**
-   * Add an `operation` to the transform that resulted in `state`.
-   *
-   * @param {Object} operation
-   * @return {Transform}
-   */
-
-  operate(operation) {
-    const { type } = operation
-    const fn = Operations[type]
-
-    if (!fn) {
-      throw new Error(`Unknown operation type: "${type}".`)
-    }
-
-    this.state = fn(this.state, operation)
-    this.operations.push(operation)
-    return this
-  }
-
-  /**
    * Apply the transform and return the new state.
    *
    * @param {Object} options
@@ -64,38 +45,38 @@ class Transform {
    */
 
   apply(options = {}) {
-    let { initialState, state, operations } = this
-    let { history, selection } = state
-    let { marks } = selection
+    let { state, operations } = this
+    let { history } = state
     let { undos, redos } = history
+
+    // If there are no operations, abort early.
+    if (!operations.length) return state
+
+    // The `isNative` flag allows for natively-handled changes to skip
+    // rerendering the editor for improved performance.
+    const isNative = !!options.isNative
 
     // Determine whether we need to create a new snapshot.
     const shouldSnapshot = options.snapshot == null
       ? this.shouldSnapshot()
       : options.snapshot
 
-    // If we should, save a snapshot into the history before transforming.
+    // Either create a new snapshot, or push the operations into the previous.
     if (shouldSnapshot) {
-      const snapshot = this.snapshot()
+      const snapshot = { operations }
       undos = undos.push(snapshot)
-      if (undos.size > 100) undos = undos.take(100)
-      redos = redos.clear()
-      history = history.merge({ undos, redos })
-      state = state.merge({ history })
+    } else {
+      const snapshot = undos.peek()
+      snapshot.operations = snapshot.operations.concat(operations)
     }
 
-    // Check whether to remove the cursor marks.
-    // if (options.snapshot !== false && initialState.selection.marks) {
-    //   selection = selection.merge({ marks: null })
-    //   state = state.merge({ selection })
-    // }
+    // Clear the redo stack and constrain the undos stack.
+    if (undos.size > 100) undos = undos.take(100)
+    redos = redos.clear()
 
-    // Apply the "isNative" flag, which is used to allow for natively-handled
-    // content changes to skip rerendering the editor for performance.
-    state = state.merge({
-      isNative: !!options.isNative
-    })
-
+    // Update the state.
+    history = history.merge({ undos, redos })
+    state = state.merge({ history, isNative })
     return state
   }
 
@@ -114,67 +95,30 @@ class Transform {
     // If there isn't a previous state, snapshot.
     if (!previous) return true
 
-    // If the only operations applied are selection operations, don't snapshot.
-    const onlySelections = operations.every(op => op.type == 'set_selection')
-    if (onlySelections) return false
+    const types = operations.map(op => op.type)
+    const prevTypes = previous.operations.map(op => op.type)
+    const edits = types.filter(type => type != 'set_selection')
+    const prevEdits = prevTypes.filter(type => type != 'set_selection')
 
-    // If the current operations aren't one of the "combinable" types, snapshot.
-    const onlyInsert = operations.every(op => op.type == 'insert_text')
-    const onlyRemove = operations.every(op => op.type == 'remove_text')
-    const onlyPrevInsert = previous.operations.every(op => op.type == 'insert_text')
-    const onlyPrevRemove = previous.operations.every(op => op.type == 'remove_text')
-    if (onlyInsert && onlyPrevInsert) return false
-    if (onlyRemove && onlyPrevRemove) return false
+    const onlySelections = types.every(type => type == 'set_selection')
+    const onlyInsert = edits.every(type => type == 'insert_text')
+    const onlyRemove = edits.every(type => type == 'remove_text')
+    const prevOnlySelections = prevTypes.every(type => type == 'set_selection')
+    const prevOnlyInsert = prevEdits.every(type => type == 'insert_text')
+    const prevOnlyRemove = prevEdits.every(type => type == 'remove_text')
+
+    // If the only operations applied are selection operations, or if the
+    // current operations are all text editing, don't snapshot.
+    if (
+      (onlySelections) ||
+      (!prevOnlySelections && onlyInsert && prevOnlyInsert) ||
+      (!prevOnlySelections && onlyRemove && prevOnlyRemove)
+    ) {
+      return false
+    }
 
     // Otherwise, snapshot.
     return true
-  }
-
-  /**
-   * Create a history-ready snapshot of the current state.
-   *
-   * @return {Object}
-   */
-
-  snapshot() {
-    let { initialState, operations } = this
-    let { document, selection } = initialState
-    return { document, selection, operations }
-  }
-
-  /**
-   * Undo to the previous state in the history.
-   *
-   * @return {State} state
-   */
-
-  undo() {
-    let { state } = this
-    let { history } = state
-    let { undos, redos } = history
-
-    // If there's no previous snapshot, return the current state.
-    let previous = undos.peek()
-    if (!previous) return state
-
-    // Remove the previous snapshot from the undo stack.
-    undos = undos.pop()
-
-    // Snapshot the current state, and move it into the redos stack.
-    let snapshot = this.snapshot()
-    redos = redos.push(snapshot)
-
-    // Return the previous state, with the updated history.
-    let { document, selection } = previous
-    history = history.merge({ undos, redos })
-    state = state.merge({
-      document,
-      selection,
-      history,
-      isNative: false
-    })
-
-    return state
   }
 
   /**
@@ -192,21 +136,58 @@ class Transform {
     let next = redos.peek()
     if (!next) return state
 
-    // Remove the next history from the redo stack.
+    // Shift the next state into the undo stack.
     redos = redos.pop()
+    undos = undos.push(next)
 
-    // Snapshot the current state, and move it into the undos stack.
-    let snapshot = this.snapshot()
-    undos = undos.push(snapshot)
+    // Replay the next operations.
+    const { operations } = next
+    operations.forEach(op => {
+      this.applyOperation(op)
+    })
 
-    // Return the next state, with the updated history.
-    let { document, selection } = next
+    // Update the state's history and force `isNative` to false.
     history = history.merge({ undos, redos })
-    state = state.merge({
-      document,
-      selection,
+    state = this.state.merge({
       history,
-      isNative: false
+      isNative: false,
+    })
+
+    return state
+  }
+
+  /**
+   * Undo the previous operations in the history.
+   *
+   * @return {State} state
+   */
+
+  undo() {
+    let { state } = this
+    let { history } = state
+    let { undos, redos } = history
+
+    // If there's no previous snapshot, return the current state.
+    let previous = undos.peek()
+    if (!previous) return state
+
+    // Shift the previous operations into the redo stack.
+    undos = undos.pop()
+    redos = redos.push(previous)
+
+    // Replay the inverse of the previous operations.
+    const operations = previous.operations.slice().reverse()
+    operations.forEach(op => {
+      op.inverse.forEach(inv => {
+        this.applyOperation(inv)
+      })
+    })
+
+    // Update the state's history and force `isNative` to false.
+    history = history.merge({ undos, redos })
+    state = this.state.merge({
+      history,
+      isNative: false,
     })
 
     return state
