+++ lib/plugins/core.js
@@ -568,7 +568,6 @@ function Plugin(options = {}) {
     return state
       .transform()
       .moveTo(selection)
-      .focus()
       .apply()
   }
