+++ lib/transforms/at-current-range.js
@@ -22,13 +22,13 @@ export function addMark(transform, mark) {
   else if (selection.marks) {
     const marks = selection.marks.add(mark)
     const sel = selection.merge({ marks })
-    return transform.setSelection(sel)
+    return transform.moveTo(sel)
   }
 
   else {
     const marks = document.getMarksAtRange(selection).add(mark)
     const sel = selection.merge({ marks })
-    return transform.setSelection(sel)
+    return transform.moveTo(sel)
   }
 }
 
@@ -81,7 +81,7 @@ export function _delete(transform) {
 
   return transform
     .deleteAtRange(selection)
-    .setSelection(after)
+    .moveTo(after)
 }
 
 /**
@@ -155,7 +155,7 @@ export function deleteBackward(transform, n = 1) {
 
   return transform
     .deleteBackwardAtRange(selection, n)
-    .setSelection(after)
+    .moveTo(after)
 }
 
 /**
@@ -214,7 +214,7 @@ export function deleteForward(transform, n = 1) {
 
   return transform
     .deleteForwardAtRange(selection, n)
-    .setSelection(after)
+    .moveTo(after)
 }
 
 /**
@@ -237,7 +237,7 @@ export function insertBlock(transform, block) {
   const text = document.getTexts().find(n => !keys.includes(n.key))
   const after = selection.collapseToEndOf(text)
 
-  return transform.setSelection(after)
+  return transform.moveTo(after)
 }
 
 /**
@@ -287,7 +287,7 @@ export function insertFragment(transform, fragment) {
       .moveForward(lastText.length)
   }
 
-  return transform.setSelection(after)
+  return transform.moveTo(after)
 }
 
 /**
@@ -319,7 +319,7 @@ export function insertInline(transform, inline) {
     after = selection.collapseToEndOf(text)
   }
 
-  return transform.setSelection(after)
+  return transform.moveTo(after)
 }
 
 /**
@@ -354,7 +354,7 @@ export function insertText(transform, text, marks) {
 
   return transform
     .insertTextAtRange(selection, text, marks)
-    .setSelection(after)
+    .moveTo(after)
 }
 
 /**
@@ -407,7 +407,7 @@ export function splitBlock(transform, depth = 1) {
   const nextNode = document.getNextText(startNode)
   const after = selection.collapseToStartOf(nextNode)
 
-  return transform.setSelection(after)
+  return transform.moveTo(after)
 }
 
 /**
@@ -436,7 +436,7 @@ export function splitInline(transform, depth = Infinity) {
     after = selection.collapseToStartOf(nextNode)
   }
 
-  return transform.setSelection(after)
+  return transform.moveTo(after)
 }
 
 /**
@@ -460,13 +460,13 @@ export function removeMark(transform, mark) {
   else if (selection.marks) {
     const marks = selection.marks.remove(mark)
     const sel = selection.merge({ marks })
-    return transform.setSelection(sel)
+    return transform.moveTo(sel)
   }
 
   else {
     const marks = document.getMarksAtRange(selection).remove(mark)
     const sel = selection.merge({ marks })
-    return transform.setSelection(sel)
+    return transform.moveTo(sel)
   }
 }
 
@@ -582,7 +582,7 @@ export function wrapInline(transform, properties) {
   }
 
   after = after.normalize(document)
-  return transform.setSelection(after)
+  return transform.moveTo(after)
 }
 
 /**
@@ -613,5 +613,5 @@ export function wrapText(transform, prefix, suffix = prefix) {
 
   return transform
     .wrapTextAtRange(selection, prefix, suffix)
-    .setSelection(after)
+    .moveTo(after)
 }
