+++ lib/transforms/by-key.js
@@ -15,29 +15,10 @@ import uid from '../utils/uid'
 
 export function addMarkByKey(transform, key, offset, length, mark) {
   mark = Normalize.mark(mark)
-
   const { state } = transform
   const { document } = state
   const path = document.getPath(key)
-
-  const inverse = {
-    type: 'remove_mark',
-    path,
-    offset,
-    length,
-    mark,
-  }
-
-  const operation = {
-    type: 'add_mark',
-    path,
-    offset,
-    length,
-    mark,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.addMarkOperation(path, offset, length, mark)
 }
 
 /**
@@ -55,21 +36,7 @@ export function insertNodeByKey(transform, key, index, node) {
   const { document } = state
   const path = document.getPath(key)
   const newPath = path.slice().push(index)
-
-  const inverse = {
-    type: 'remove_node',
-    path: newPath,
-  }
-
-  const operation = {
-    type: 'insert_node',
-    path,
-    index,
-    node,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.insertNodeOperation(path, index, node)
 }
 
 /**
@@ -87,23 +54,7 @@ export function insertTextByKey(transform, key, offset, text, marks) {
   const { state } = transform
   const { document } = state
   const path = document.getPath(key)
-
-  const inverse = {
-    type: 'remove_text',
-    path,
-    offset,
-    length: text.length,
-  }
-
-  const operation = {
-    type: 'insert_text',
-    path,
-    offset,
-    text,
-    marks,
-  }
-
-  return transform.operate(operation)
+  return transform.insertTextOperation(path, offset, text, marks)
 }
 
 /**
@@ -119,30 +70,9 @@ export function insertTextByKey(transform, key, offset, text, marks) {
 export function moveNodeByKey(transform, key, newKey, newIndex) {
   const { state } = transform
   const { document } = state
-  const node = document.assertDescendant(key)
   const path = document.getPath(key)
-  const parent = document.getParent(key)
-  const parentPath = path.slice(0, -1)
-  const parentIndex = path[path.length - 1]
   const newPath = document.getPath(newKey)
-  const nodePath = newPath.slice().concat([newIndex])
-
-  const inverse = {
-    type: 'move_node',
-    path: nodePath,
-    newPath: parentPath,
-    newIndex: parentIndex,
-  }
-
-  const operation = {
-    type: 'move_node',
-    path,
-    newPath,
-    newIndex,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.moveNodeOperation(path, newPath, newIndex)
 }
 
 /**
@@ -158,29 +88,10 @@ export function moveNodeByKey(transform, key, newKey, newIndex) {
 
 export function removeMarkByKey(transform, key, offset, length, mark) {
   mark = Normalize.mark(mark)
-
   const { state } = transform
   const { document } = state
   const path = document.getPath(key)
-
-  const inverse = {
-    type: 'add_mark',
-    path,
-    offset,
-    length,
-    mark,
-  }
-
-  const operation = {
-    type: 'remove_mark',
-    path,
-    offset,
-    length,
-    mark,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.removeMarkOperation(path, offset, length, mark)
 }
 
 /**
@@ -194,25 +105,8 @@ export function removeMarkByKey(transform, key, offset, length, mark) {
 export function removeNodeByKey(transform, key) {
   const { state } = transform
   const { document } = state
-  const node = document.assertDescendant(key)
   const path = document.getPath(key)
-  const parentPath = path.slice(0, -1)
-  const parentIndex = path.slice(-1)
-
-  const inverse = {
-    type: 'insert_node',
-    path: parentPath,
-    index: parentIndex,
-    node,
-  }
-
-  const operation = {
-    type: 'remove_node',
-    path,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.removeNodeOperation(path)
 }
 
 /**
@@ -229,19 +123,7 @@ export function removeTextByKey(transform, key, offset, length) {
   const { state } = transform
   const { document } = state
   const path = document.getPath(key)
-
-  // TODO!
-  const inverse = {}
-
-  const operation = {
-    type: 'remove_text',
-    path,
-    offset,
-    length,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.removeTextOperation(path, offset, length)
 }
 
 /**
@@ -258,36 +140,10 @@ export function removeTextByKey(transform, key, offset, length) {
 export function setMarkByKey(transform, key, offset, length, mark, properties) {
   mark = Normalize.mark(mark)
   properties = Normalize.markProperties(properties)
-
   const { state } = transform
   const { document } = state
   const path = document.getPath(key)
-  const prevProps = {}
-
-  for (const k in properties) {
-    prevProps[k] = mark[k]
-  }
-
-  const inverse = {
-    type: 'set_mark',
-    path,
-    offset,
-    length,
-    mark,
-    properties: prevProps,
-  }
-
-  const operation = {
-    type: 'set_mark',
-    path,
-    offset,
-    length,
-    mark,
-    properties,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.setMarkOperation(path, offset, length, mark, properties)
 }
 
 /**
@@ -301,31 +157,11 @@ export function setMarkByKey(transform, key, offset, length, mark, properties) {
 
 export function setNodeByKey(transform, key, properties) {
   properties = Normalize.nodeProperties(properties)
-
   const { state } = transform
   const { document } = state
   const node = document.assertDescendant(key)
   const path = document.getPath(key)
-  const prevProps = {}
-
-  for (const k in properties) {
-    prevProps[k] = node[k]
-  }
-
-  const inverse = {
-    type: 'set_node',
-    path,
-    properties: prevProps
-  }
-
-  const operation = {
-    type: 'set_node',
-    path,
-    properties,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.setNodeOperation(path, properties)
 }
 
 /**
@@ -341,17 +177,5 @@ export function splitNodeByKey(transform, key, offset) {
   const { state } = transform
   const { document } = state
   const path = document.getPath(key)
-
-  // TODO!
-  const inverse = {}
-
-  const operation = {
-    type: 'split_node',
-    path,
-    offset,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.splitNodeOperation(path, offset)
 }
-
