+++ lib/transforms/index.js
@@ -1,5 +1,31 @@
 
 /**
+ * Apply operation.
+ */
+
+import {
+  applyOperation,
+} from './apply-operation'
+
+/**
+ * Operations.
+ */
+
+import {
+  addMarkOperation,
+  insertNodeOperation,
+  insertTextOperation,
+  moveNodeOperation,
+  removeMarkOperation,
+  removeNodeOperation,
+  removeTextOperation,
+  setMarkOperation,
+  setNodeOperation,
+  setSelectionOperation,
+  splitNodeOperation,
+} from './operations'
+
+/**
  * At range.
  */
 
@@ -98,7 +124,6 @@ import {
   moveTo,
   moveToOffsets,
   moveToRangeOf,
-  setSelection,
 } from './on-selection'
 
 /**
@@ -117,6 +142,28 @@ import {
  */
 
 export default {
+
+  /**
+   * Apply operation.
+   */
+
+  applyOperation,
+
+  /**
+   * Operations.
+   */
+
+  addMarkOperation,
+  insertNodeOperation,
+  insertTextOperation,
+  moveNodeOperation,
+  removeMarkOperation,
+  removeNodeOperation,
+  removeTextOperation,
+  setMarkOperation,
+  setNodeOperation,
+  setSelectionOperation,
+  splitNodeOperation,
 
   /**
    * At range.
@@ -210,7 +257,6 @@ export default {
   moveTo,
   moveToOffsets,
   moveToRangeOf,
-  setSelection,
 
   /**
    * Normalize.
