+++ lib/transforms/on-selection.js
@@ -1,7 +1,4 @@
 
-import Normalize from '../utils/normalize'
-import Selection from '../models/selection'
-
 /**
  * Blur the selection.
  *
@@ -13,7 +10,7 @@ export function blur(transform) {
   const { state } = transform
   const { selection } = state
   const sel = selection.blur()
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -27,7 +24,7 @@ export function collapseToAnchor(transform) {
   const { state } = transform
   const { selection } = state
   const sel = selection.collapseToAnchor()
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -41,7 +38,7 @@ export function collapseToEnd(transform) {
   const { state } = transform
   const { selection } = state
   const sel = selection.collapseToEnd()
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -55,7 +52,7 @@ export function collapseToFocus(transform) {
   const { state } = transform
   const { selection } = state
   const sel = selection.collapseToFocus()
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -69,7 +66,7 @@ export function collapseToStart(transform) {
   const { state } = transform
   const { selection } = state
   const sel = selection.collapseToStart()
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -84,7 +81,7 @@ export function collapseToEndOf(transform, node) {
   const { state } = transform
   const { selection } = state
   const sel = selection.collapseToEndOf(node)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -103,7 +100,7 @@ export function collapseToEndOfNextBlock(transform) {
   if (!next) return transform
 
   const sel = selection.collapseToEndOf(next)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -122,7 +119,7 @@ export function collapseToEndOfNextText(transform) {
   if (!next) return transform
 
   const sel = selection.collapseToEndOf(next)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -141,7 +138,7 @@ export function collapseToEndOfPreviousBlock(transform) {
   if (!previous) return transform
 
   const sel = selection.collapseToEndOf(previous)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -160,7 +157,7 @@ export function collapseToEndOfPreviousText(transform) {
   if (!previous) return transform
 
   const sel = selection.collapseToEndOf(previous)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -175,7 +172,7 @@ export function collapseToStartOf(transform, node) {
   const { state } = transform
   const { selection } = state
   const sel = selection.collapseToStartOf(node)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -194,7 +191,7 @@ export function collapseToStartOfNextBlock(transform) {
   if (!next) return transform
 
   const sel = selection.collapseToStartOf(next)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -213,7 +210,7 @@ export function collapseToStartOfNextText(transform) {
   if (!next) return transform
 
   const sel = selection.collapseToStartOf(next)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -232,7 +229,7 @@ export function collapseToStartOfPreviousBlock(transform) {
   if (!previous) return transform
 
   const sel = selection.collapseToStartOf(previous)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -251,7 +248,7 @@ export function collapseToStartOfPreviousText(transform) {
   if (!previous) return transform
 
   const sel = selection.collapseToStartOf(previous)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -266,7 +263,7 @@ export function extendBackward(transform, n) {
   const { state } = transform
   const { document, selection } = state
   const sel = selection.extendBackward(n).normalize(document)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -281,7 +278,7 @@ export function extendForward(transform, n) {
   const { state } = transform
   const { document, selection } = state
   const sel = selection.extendForward(n).normalize(document)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -296,7 +293,7 @@ export function extendToEndOf(transform, node) {
   const { state } = transform
   const { document, selection } = state
   const sel = selection.extendToEndOf(node).normalize(document)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -311,7 +308,7 @@ export function extendToStartOf(transform, node) {
   const { state } = transform
   const { document, selection } = state
   const sel = selection.extendToStartOf(node).normalize(document)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -325,7 +322,7 @@ export function focus(transform) {
   const { state } = transform
   const { selection } = state
   const sel = selection.focus()
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -340,7 +337,7 @@ export function moveBackward(transform, n) {
   const { state } = transform
   const { document, selection } = state
   const sel = selection.moveBackward(n).normalize(document)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -355,7 +352,7 @@ export function moveForward(transform, n) {
   const { state } = transform
   const { document, selection } = state
   const sel = selection.moveForward(n).normalize(document)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -367,11 +364,7 @@ export function moveForward(transform, n) {
  */
 
 export function moveTo(transform, properties) {
-  properties = Normalize.selection(properties)
-  const { state } = transform
-  const { document, selection } = state
-  const sel = selection.merge(properties).normalize(document)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(properties)
 }
 
 /**
@@ -387,7 +380,7 @@ export function moveToOffsets(transform, anchor, fokus) {
   const { state } = transform
   const { document, selection } = state
   const sel = selection.moveToOffsets(anchor, fokus)
-  return transform.setSelection(sel)
+  return transform.setSelectionOperation(sel)
 }
 
 /**
@@ -403,42 +396,5 @@ export function moveToRangeOf(transform, start, end) {
   const { state } = transform
   const { document, selection } = state
   const sel = selection.moveToRangeOf(start, end).normalize(document)
-  return transform.setSelection(sel)
-}
-
-/**
- * Set the selection to a new `selection`.
- *
- * @param {Transform} transform
- * @param {Mixed} selection
- * @return {Transform}
- */
-
-export function setSelection(transform, properties) {
-  properties = Normalize.selectionProperties(properties)
-
-  const { state } = transform
-  const { selection } = state
-  const prevProps = {}
-
-  if (properties.marks == selection.marks) {
-    properties.marks = null
-  }
-
-  for (const k in properties) {
-    prevProps[k] = selection[k]
-  }
-
-  const inverse = {
-    type: 'set_selection',
-    properties: prevProps
-  }
-
-  const operation = {
-    type: 'set_selection',
-    properties,
-    inverse,
-  }
-
-  return transform.operate(operation)
+  return transform.setSelectionOperation(sel)
 }
