+++ lib/transforms/operations.js
@@ -1,300 +1,375 @@
 
-import Debug from 'debug'
+import Normalize from '../utils/normalize'
 import uid from '../utils/uid'
-
-/**
- * Debug.
- *
- * @type {Function}
- */
-
-const debug = Debug('slate:operation')
 
 /**
  * Add mark to text at `offset` and `length` in node by `path`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @param {Number} offset
+ * @param {Number} length
+ * @param {Mixed} mark
+ * @return {Transform}
  */
 
-function addMark(state, operation) {
-  debug('add_mark', operation)
-  const { path, offset, length, mark } = operation
-  let { document } = state
-  let node = document.assertPath(path)
-  node = node.addMark(offset, length, mark)
-  document = document.updateDescendant(node)
-  state = state.merge({ document })
-  return state
+export function addMarkOperation(transform, path, offset, length, mark) {
+  const inverse = [{
+    type: 'remove_mark',
+    path,
+    offset,
+    length,
+    mark,
+  }]
+
+  const operation = {
+    type: 'add_mark',
+    path,
+    offset,
+    length,
+    mark,
+    inverse,
+  }
+
+  return transform.applyOperation(operation)
 }
 
 /**
  * Insert a `node` at `index` in a node by `path`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @param {Number} index
+ * @param {Node} node
+ * @return {Transform}
  */
 
-function insertNode(state, operation) {
-  debug('insert_node', operation)
-  const { path, index, node } = operation
-  let { document } = state
-  let parent = document.assertPath(path)
-  const isParent = document == parent
-  const nodes = parent.nodes.splice(index, 0, node)
-  parent = parent.merge({ nodes })
-  document = isParent ? parent : document.updateDescendant(parent)
-  state = state.merge({ document })
-  return state
+export function insertNodeOperation(transform, path, index, node) {
+  const inversePath = path.slice().concat([index])
+  const inverse = [{
+    type: 'remove_node',
+    path: inversePath,
+  }]
+
+  const operation = {
+    type: 'insert_node',
+    path,
+    index,
+    node,
+    inverse,
+  }
+
+  return transform.applyOperation(operation)
 }
 
 /**
  * Insert `text` at `offset` in node by `path`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @param {Number} offset
+ * @param {String} text
+ * @param {Set} marks (optional)
+ * @return {Transform}
  */
 
-function insertText(state, operation) {
-  debug('insert_text', operation)
-  const { path, offset, text, marks } = operation
-  let { document } = state
-  let node = document.assertPath(path)
-  node = node.insertText(offset, text, marks)
-  document = document.updateDescendant(node)
-  state = state.merge({ document })
-  return state
+export function insertTextOperation(transform, path, offset, text, marks) {
+  const inverseLength = text.length
+  const inverse = [{
+    type: 'remove_text',
+    path,
+    offset,
+    length: inverseLength,
+  }]
+
+  const operation = {
+    type: 'insert_text',
+    path,
+    offset,
+    text,
+    marks,
+    inverse,
+  }
+
+  return transform.applyOperation(operation)
 }
 
 /**
- * Move a node by `path` to a new parent by `path` and `index`.
+ * Move a node by `path` to a `newPath` and `newIndex`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @param {Array} newPath
+ * @param {Number} newIndex
+ * @return {Transform}
  */
 
-function moveNode(state, operation) {
-  debug('move_node', operation)
-  const { path, newPath, newIndex } = operation
-  let { document } = state
-  const node = document.assertPath(path)
+export function moveNodeOperation(transform, path, newPath, newIndex) {
+  const { state } = transform
+  const { document } = state
+  const parentPath = path.slice(0, -1)
+  const parentIndex = path[path.length - 1]
+  const inversePath = newPath.slice().concat([newIndex])
 
-  let parent = document.getParent(node)
-  const isParent = document == parent
-  const index = parent.nodes.indexOf(node)
-  parent = parent.removeNode(index)
-  document = isParent ? parent : document.updateDescendant(parent)
+  const inverse = [{
+    type: 'move_node',
+    path: inversePath,
+    newPath: parentPath,
+    newIndex: parentIndex,
+  }]
 
-  let target = document.assertPath(newPath)
-  const isTarget = document == target
-  target = target.insertNode(newIndex, node)
-  document = isTarget ? target : document.updateDescendant(target)
+  const operation = {
+    type: 'move_node',
+    path,
+    newPath,
+    newIndex,
+    inverse,
+  }
 
-  state = state.merge({ document })
-  return state
+  return transform.applyOperation(operation)
 }
 
 /**
  * Remove mark from text at `offset` and `length` in node by `path`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @param {Number} offset
+ * @param {Number} length
+ * @param {Mark} mark
+ * @return {Transform}
  */
 
-function removeMark(state, operation) {
-  debug('remove_mark', operation)
-  const { path, offset, length, mark } = operation
-  let { document } = state
-  let node = document.assertPath(path)
-  node = node.removeMark(offset, length, mark)
-  document = document.updateDescendant(node)
-  state = state.merge({ document })
-  return state
+export function removeMarkOperation(transform, path, offset, length, mark) {
+  const inverse = [{
+    type: 'add_mark',
+    path,
+    offset,
+    length,
+    mark,
+  }]
+
+  const operation = {
+    type: 'remove_mark',
+    path,
+    offset,
+    length,
+    mark,
+    inverse,
+  }
+
+  return transform.applyOperation(operation)
 }
 
 /**
  * Remove a node by `path`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @return {Transform}
  */
 
-function removeNode(state, operation) {
-  debug('remove_node', operation)
-  const { path } = operation
-  let { document } = state
+export function removeNodeOperation(transform, path) {
+  const { state } = transform
+  const { document } = state
   const node = document.assertPath(path)
-  let parent = document.getParent(node)
-  const index = parent.nodes.indexOf(node)
-  const isParent = document == parent
-  parent = parent.removeNode(index)
-  document = isParent ? parent : document.updateDescendant(parent)
-  document = document.normalize()
-  state = state.merge({ document })
-  return state
+  const inversePath = path.slice(0, -1)
+  const inverseIndex = path.slice(-1)
+
+  const inverse = [{
+    type: 'insert_node',
+    path: inversePath,
+    index: inverseIndex,
+    node,
+  }]
+
+  const operation = {
+    type: 'remove_node',
+    path,
+    inverse,
+  }
+
+  return transform.applyOperation(operation)
 }
 
 /**
  * Remove text at `offset` and `length` in node by `path`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @param {Number} offset
+ * @param {Number} length
+ * @return {Transform}
  */
 
-function removeText(state, operation) {
-  debug('remove_text', operation)
-  const { path, offset, length } = operation
-  let { document } = state
-  let node = document.assertPath(path)
-  node = node.removeText(offset, length)
-  document = document.updateDescendant(node)
-  document = document.normalize()
-  state = state.merge({ document })
-  return state
+export function removeTextOperation(transform, path, offset, length) {
+  const { state } = transform
+  const { document } = state
+  const node = document.assertPath(path)
+  const ranges = node.getRanges()
+  const inverse = []
+
+  ranges.reduce((start, range) => {
+    const { text, marks } = range
+    const end = start + text.length
+    if (start > offset + length) return
+    if (end <= offset) return
+
+    const endOffset = Math.min(end, offset + length)
+    const string = text.slice(offset, endOffset)
+
+    inverse.push({
+      type: 'insert_text',
+      path,
+      offset,
+      text: string,
+      marks,
+    })
+
+    return end
+  }, 0)
+
+  const operation = {
+    type: 'remove_text',
+    path,
+    offset,
+    length,
+    inverse,
+  }
+
+  return transform.applyOperation(operation)
 }
 
 /**
  * Set `properties` on mark on text at `offset` and `length` in node by `path`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @param {Number} offset
+ * @param {Number} length
+ * @param {Mark} mark
+ * @return {Transform}
  */
 
-function setMark(state, operation) {
-  debug('set_mark', operation)
-  const { path, offset, length, mark, properties } = operation
-  let { document } = state
-  let node = document.assertPath(path)
-  node = node.updateMark(offset, length, mark, properties)
-  document = document.updateDescendant(node)
-  state = state.merge({ document })
-  return state
+export function setMarkOperation(transform, path, offset, length, mark, properties) {
+  const inverseProps = {}
+
+  for (const k in properties) {
+    inverseProps[k] = mark[k]
+  }
+
+  const inverse = [{
+    type: 'set_mark',
+    path,
+    offset,
+    length,
+    mark,
+    properties: inverseProps,
+  }]
+
+  const operation = {
+    type: 'set_mark',
+    path,
+    offset,
+    length,
+    mark,
+    properties,
+    inverse,
+  }
+
+  return transform.applyOperation(operation)
 }
 
 /**
  * Set `properties` on a node by `path`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @param {Object || String} properties
+ * @return {Transform}
  */
 
-function setNode(state, operation) {
-  debug('set_node', operation)
-  const { path, properties } = operation
-  let { document } = state
-  let node = document.assertPath(path)
-  node = node.merge(properties)
-  document = document.updateDescendant(node)
-  document = document.normalize()
-  state = state.merge({ document })
-  return state
+export function setNodeOperation(transform, path, properties) {
+  const { state } = transform
+  const { document } = state
+  const node = document.assertPath(path)
+  const inverseProps = {}
+
+  for (const k in properties) {
+    inverseProps[k] = node[k]
+  }
+
+  const inverse = [{
+    type: 'set_node',
+    path,
+    properties: inverseProps
+  }]
+
+  const operation = {
+    type: 'set_node',
+    path,
+    properties,
+    inverse,
+  }
+
+  return transform.applyOperation(operation)
 }
 
 /**
- * Set `properties` on the selection.
+ * Set the selection to a new `selection`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Mixed} selection
+ * @return {Transform}
  */
 
-function setSelection(state, operation) {
-  debug('set_selection', operation)
-  let { properties } = operation
-  let { selection } = state
-  selection = selection.merge(properties)
-  state = state.merge({ selection })
-  return state
+export function setSelectionOperation(transform, properties) {
+  properties = Normalize.selectionProperties(properties)
+
+  const { state } = transform
+  const { selection } = state
+  const prevProps = {}
+
+  if (properties.marks == selection.marks) {
+    properties.marks = null
+  }
+
+  for (const k in properties) {
+    prevProps[k] = selection[k]
+  }
+
+  const inverse = [{
+    type: 'set_selection',
+    properties: prevProps
+  }]
+
+  const operation = {
+    type: 'set_selection',
+    properties,
+    inverse,
+  }
+
+  return transform.applyOperation(operation)
 }
 
 /**
  * Split a node by `path` at `offset`.
  *
- * @param {State} state
- * @param {Object} operation
- * @return {State}
+ * @param {Transform} transform
+ * @param {Array} path
+ * @param {Number} offset
+ * @return {Transform}
  */
 
-function splitNode(state, operation) {
-  debug('split_node', operation)
-  const { path, offset } = operation
-  let { document } = state
-  let node = document.assertPath(path)
-  let parent = document.getParent(node)
-  const isParent = document == parent
-  const index = parent.nodes.indexOf(node)
+export function splitNodeOperation(transform, path, offset) {
+  const inverse = []
 
-  let child = node
-  let one
-  let two
-
-  if (node.kind != 'text') {
-    child = node.getTextAtOffset(offset)
+  const operation = {
+    type: 'split_node',
+    path,
+    offset,
+    inverse,
   }
 
-  while (child && child != parent) {
-    if (child.kind == 'text') {
-      const i = node.kind == 'text' ? offset : offset - node.getOffset(child)
-      const { characters } = child
-      const oneChars = characters.take(i)
-      const twoChars = characters.skip(i)
-      one = child.merge({ characters: oneChars })
-      two = child.merge({ characters: twoChars, key: uid() })
-    }
-
-    else {
-      const { nodes } = child
-      const oneNodes = nodes.takeUntil(n => n.key == one.key).push(one)
-      const twoNodes = nodes.skipUntil(n => n.key == one.key).rest().unshift(two)
-      one = child.merge({ nodes: oneNodes })
-      two = child.merge({ nodes: twoNodes, key: uid() })
-    }
-
-    child = document.getParent(child)
-  }
-
-  parent = parent.removeNode(index)
-  parent = parent.insertNode(index, two)
-  parent = parent.insertNode(index, one)
-  document = isParent ? parent : document.updateDescendant(parent)
-  state = state.merge({ document })
-  return state
-}
-
-/**
- * Export.
- *
- * @type {Object}
- */
-
-export default {
-
-  // Text operations.
-  insert_text: insertText,
-  remove_text: removeText,
-
-  // Mark operations.
-  add_mark: addMark,
-  remove_mark: removeMark,
-  set_mark: setMark,
-
-  // Node operations.
-  insert_node: insertNode,
-  move_node: moveNode,
-  remove_node: removeNode,
-  set_node: setNode,
-  split_node: splitNode,
-
-  // Selection operations.
-  set_selection: setSelection,
-
+  return transform.applyOperation(operation)
 }
